next to work on the interface

+++ cpp_tests/netology/race_simulator/lib/LandVehicle.h
@@ -10,7 +10,7 @@
 
 class LandVehicle : public Vehicle {
 public:
-    LandVehicle(std::string n, int s, int r) : Vehicle(std::move(n), s), restTime(r) {}
+    LandVehicle(std::string n, float s, int r) : Vehicle(std::move(n), s), restTime(r) {}
 protected:
     int restTime;
 };
